requirements.txt
Flask
PyPDF2
python-dotenv
langchain
chromadb
faiss-cpu
langchain-fireworks
langchain-community